Extract string handling methods from DrawsBoxes
