+++ src/Themes/Default/Concerns/DrawsBoxes.php
@@ -6,6 +6,8 @@ use Laravel\Prompts\Prompt;
 
 trait DrawsBoxes
 {
+    use HandlesStrings;
+
     protected int $minWidth = 60;
 
     /**
@@ -54,45 +56,5 @@ trait DrawsBoxes
         ).($info ? " {$info} " : '').'┘'));
 
         return $this;
-    }
-
-    /**
-     * Get the length of the longest line.
-     *
-     * @param  array<string>  $lines
-     */
-    protected function longest(array $lines, int $padding = 0): int
-    {
-        return max(
-            $this->minWidth,
-            collect($lines)
-                ->map(fn ($line) => mb_strwidth($this->stripEscapeSequences($line)) + $padding)
-                ->max()
-        );
-    }
-
-    /**
-     * Pad text ignoring ANSI escape sequences.
-     */
-    protected function pad(string $text, int $length, $char = ' '): string
-    {
-        $rightPadding = str_repeat($char, max(0, $length - mb_strwidth($this->stripEscapeSequences($text))));
-
-        return "{$text}{$rightPadding}";
-    }
-
-    /**
-     * Strip ANSI escape sequences from the given text.
-     */
-    protected function stripEscapeSequences(string $text): string
-    {
-        // Strip ANSI escape sequences.
-        $text = preg_replace("/\e[^m]*m/", '', $text);
-
-        // Strip Symfony named style tags.
-        $text = preg_replace("/<(info|comment|question|error)>(.*?)<\/\\1>/", '$2', $text);
-
-        // Strip Symfony inline style tags.
-        return preg_replace("/<(?:(?:[fb]g|options)=[a-z,;]+)+>(.*?)<\/>/i", '$1', $text);
     }
 }
